Fixed rotations of player, changes particles, Kelp growth
Fixed the random rotation of player when it collides with other things.
More improvements on particles, and did the animation and grow of Kelp.
Did the trash count as well, but for some reason it's counting 2 trashes
when it gets. Im using trigger enter to count the trash. For now im just
getting the half of the total.